requirements.txt
jax>=0.2.26
jaxlib>=0.1.74
gymnax
evosax
distrax
optax
flax
numpy
brax
wandb
flashbax
navix
PyYAML
matplotlib